requirements.txt
pika==1.2.0
rfc3161ng==2.1.3
pyyaml==5.4.1
cryptography==3.4.7
pyopenssl==20.0.1
apscheduler==3.7.0
mysql-connector-python==8.0.24
paramiko==2.6.0
requests==2.22.0
python-logging-rabbitmq==2.0.0  
